+++ messaging-core/src/main/java/com/nabla/sdk/core/data/logger/HttpLoggingInterceptorFactory.kt
@@ -2,16 +2,17 @@ package com.nabla.sdk.core.data.logger
 
 import com.nabla.sdk.core.data.file.FileService
 import com.nabla.sdk.core.domain.boundary.Logger
+import com.nabla.sdk.core.domain.boundary.Logger.Companion.HTTP_DOMAIN
 import okhttp3.Interceptor
 import okhttp3.logging.HttpLoggingInterceptor
 
 internal object HttpLoggingInterceptorFactory {
     fun make(logger: Logger): Interceptor {
         val bodyLogger = HttpLoggingInterceptor { message ->
-            logger.debug(message, tag = Logger.asSdkTag("Http"))
+            logger.debug(domain = HTTP_DOMAIN, message = message)
         }.apply { setLevel(HttpLoggingInterceptor.Level.BODY) }
         val headerLogger = HttpLoggingInterceptor { message ->
-            logger.debug(message, tag = Logger.asSdkTag("Http"))
+            logger.debug(domain = HTTP_DOMAIN, message = message)
         }.apply { setLevel(HttpLoggingInterceptor.Level.HEADERS) }
 
         return Interceptor { chain ->