+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/boundary/Logger.kt
@@ -1,16 +1,19 @@
 package com.nabla.sdk.core.domain.boundary
 
+/**
+ * Abstraction of the logging layer of the SDK. You can implement this interface yourself to get
+ * access to SDK logs and errors in your application or you can use the default implementation
+ * provided by the SDK that just outputs logs in the logcat.
+ */
 public interface Logger {
-    public fun debug(message: String, error: Throwable? = null, tag: String = DEFAULT_TAG)
-    public fun info(message: String, error: Throwable? = null, tag: String = DEFAULT_TAG)
-    public fun warn(message: String, error: Throwable? = null, tag: String = DEFAULT_TAG)
-    public fun error(message: String, error: Throwable? = null, tag: String = DEFAULT_TAG)
+    public fun debug(message: String, error: Throwable? = null, domain: String)
+    public fun info(message: String, error: Throwable? = null, domain: String? = null)
+    public fun warn(message: String, error: Throwable? = null, domain: String? = null)
+    public fun error(message: String, error: Throwable? = null, domain: String? = null)
 
     public companion object {
-        private const val DEFAULT_TAG = "Nabla-SDK"
-        internal const val AUTH_TAG = "$DEFAULT_TAG-Auth"
-        public fun asSdkTag(domain: String): String {
-            return "$DEFAULT_TAG-${domain.replaceFirstChar { it.uppercaseChar() }}"
-        }
+        internal const val AUTH_DOMAIN = "Auth"
+        internal const val HTTP_DOMAIN = "Http"
+        internal const val GQL_DOMAIN = "Gql"
     }
 }